+++ src/main.js
@@ -21,6 +21,7 @@ import { renderAutomation } from './views/automation.js';
 import { renderAutomationGuide } from './views/automationGuide.js';
 import { renderAutoOrders } from './views/autoOrders.js';
 import { renderCronStatus } from './views/cronStatus.js';
+import { renderOptionsPlaybook } from './views/optionsPlaybook.js';
 import { onForegroundMessage, isFCMSupported } from './data/messaging.js';
 
 // Theme + font-size attributes are pre-set by an inline script in index.html
@@ -75,6 +76,7 @@ function mountAppShell() {
   route('automation-guide', renderAutomationGuide);
   route('auto-orders',      renderAutoOrders);
   route('cron-status',      renderCronStatus);
+  route('options-playbook', renderOptionsPlaybook);
   defaultRoute('dashboard');
   const dispatch = start(main);